refactor: move inventory templates
